Move schema populator to separate class
